changed database from double to int and removed complicated promises

+++ bamazonDB.sql
@@ -6,7 +6,7 @@ CREATE TABLE products (
 	id INT AUTO_INCREMENT,
     item VARCHAR (200) NOT NULL,
     department VARCHAR (200) NOT NULL,
-    price DOUBLE (10) NOT NULL,
+    price INT (10) NOT NULL,
     stock INT (10) NOT NULL,
     PRIMARY KEY (id)
 );
